services/upload/requirements.txt
fastapi
uvicorn
boto3
python-multipart
python-docx
langchain
langchain-community
# langchain-groq
langchain-openai
openai
docxtpl